--- README.md
@@ -1,4 +1,0 @@
-# Biblios.DSCE
-A library-management-system mini-project done as a part of academic requirement.
-Front-end has been designed using HTML, CSS and JavaScript.
-Back-end requirements have been handled by PHP. MySQL database has been used here.
