+++ logout_reader.php
@@ -9,6 +9,6 @@ $_SESSION = array();
 session_destroy();
  
 // Redirect to login page
-header("location: login_reader.php");
+header("location: logintest.php");
 exit;
 ?>
