+++ modify_book.php
@@ -45,7 +45,7 @@
 		} else {
 			echo "Error: " . $updateBook_Que . "<br>" . $link->error;
 		}
-        echo "<script>alert('Book modified successfully');document.location='bookCatalogue.php'</script>";
+        echo "<script>alert('Book modified successfully');document.location='bookcatalogue.php'</script>";
     }
 
     if(($_SERVER["REQUEST_METHOD"]=="POST") && isset($_POST["delete_book"])) {
@@ -61,7 +61,12 @@
 		} else {
 			echo "Error: " . $deleteBook_Que . "<br>" . $link->error;
 		}
-        echo "<script>alert('Book deleted successfully');document.location='bookCatalogue.php'</script>";
+        echo "<script>alert('Book deleted successfully');document.location='bookcatalogue.php'</script>";
+    }
+
+    if(($_SERVER["REQUEST_METHOD"]=="POST") && isset($_POST["cancel"])) {
+        header("location: bookcatalogue.php");
+        exit;
     }
 
     if($bookID) {
@@ -76,16 +81,85 @@
                 while($row = $result->fetch_assoc())
 						{
                 ?>
-
-                    <html>
-                        <head>
-                            <title>
-                                Modify Book
-                            </title>
-                        </head>
-                        <body>
-                            <form action="<?php echo htmlspecialchars($_SERVER["PHP_SELF"]);?>" method="post">
-                                <table>
+				
+				<html>
+				<head>
+				<title>Biblio@DSCE</title>
+					
+					<script src="https://kit.fontawesome.com/5d3eee0a99.js" crossorigin="anonymous"></script>
+					
+					<meta charset="utf-8">
+					<meta name="viewport" content="width=device-width, initial-scale=1">
+					<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+					
+					
+					<link rel="stylesheet" type="text/css" href="homeu.css">
+					<style>
+						.page-content td{
+							vertical-align:middle;
+						}
+					</style>
+				</head>
+				<body>
+				<div class="vertical-nav bg-dark text-light" id="sidebar">
+			<div class="py-4 px-3 mb-4 bg-dark text-light">
+				<div class="media d-flex align-item-center">
+					<img src="avatar1.png" alt="user image" width="80" height="80" class="mr-3 rounded-circle img-thumbnail shadow-sm">
+					<div class="media-body">
+					<h4 class="mt-3"> <?php echo htmlspecialchars($_SESSION["username"]); ?> </h4>
+					<p class="font-weight-normal text-muted mb-0">ADMIN</p>
+					</div>
+				</div>
+			</div>
+			
+			<p class="text-gray font-weight-bold text-uppercase px-3 small pb-4 mb-0">Dashboard</p>
+			<ul class="nav flex-column bg-white mb-0">
+				<li class="nav-item">
+					<a href="welcome_staff.php" class="nav-link active bg-dark text-light"><i class="fa fa-th-large mr-3 text-primary fa-fw"></i>home</a>
+				</li>
+				<li class="nav-item">
+					<a href="staffsearch.php" class="nav-link bg-dark text-light"><i class="fas fa-search mr-3 text-primary fa-fw"></i>search book</a>
+				</li>
+				<li class="nav-item">
+					<a href="newbook.php" class="nav-link bg-dark text-light"><i class="fas fa-plus-square mr-3 text-primary fa-fw"></i>add book</a>
+				</li>
+				<li class="nav-item">
+					<a href="bookcatalogue.php" class="nav-link bg-dark text-dark" id="highlight"><i class="fas fa-book-reader mr-3 text-primary fa-fw"></i>Book Catalogue</a>
+				</li>
+				<li class="nav-item">
+					<a href="staffborrow.php" class="nav-link bg-dark text-light"><i class="fas fa-book-open mr-3 text-primary fa-fw"></i>issue book</a>
+				</li>
+				<li class="nav-item">
+					<a href="outstandingpage.php" class="nav-link bg-dark text-light"><i class="fas fa-book mr-3 text-primary fa-fw"></i>outstanding books</a>
+				</li>
+				<li class="nav-item">
+					<a href="newUserReg.php" class="nav-link bg-dark text-light"><i class="fas fa-user-plus mr-3 text-primary fa-fw"></i>register new user</a>
+				</li>
+			</ul>
+			
+			<p class="text-gray font-weight-bold text-uppercase px-3 small py-4 mb-0">Charts</p>
+			<ul class="nav flex-column bg-white mb-0">
+				<li class="nav-item">
+					<a href="#" class="nav-link bg-dark text-light"><i class="fas fa-clipboard mr-3 text-primary fa-fw"></i>report</a>
+				</li>
+				<li class="nav-item">
+					<a href="#" class="nav-link bg-dark text-light"><i class="fas fa-chart-bar mr-3 text-primary fa-fw"></i>statistics</a>
+				</li>
+			</ul>
+			
+			<p class="text-gray font-weight-bold text-uppercase px-3 small py-4 mb-0"></p>
+			<ul class="nav flex-column bg-white mb-0">
+				<li class="nav-item">
+					<a href="logout_staff.php" class="nav-link bg-dark text-light"><i class="fas fa-sign-out-alt mr-3 text-primary fa-fw"></i>logout</a>
+				</li>
+			</ul>
+		
+		</div>
+		
+		<div class="page-content p-5" id="content">
+			<h1>Modify Books</h1></br>
+			<form action="<?php echo htmlspecialchars($_SERVER["PHP_SELF"]);?>" method="post">
+                                <table class="table">
                                     <tr>
                                         <td><label for="bookID"> Book ID : </label></td>
                                         <td><input id="bookID" type="text" name="id_txt" size="50" value="<?php echo $row['book_ID']?>" disabled/></td>
@@ -120,17 +194,14 @@
                                     </tr>
                                 </table>
                             <br/>
-                            <input type="submit" name="modify_book" value="Save Changes"/>
-                            <input type="submit" name="delete_book" value="Delete Book"/>
-                            <input type="submit" name="cancel" value="Cancel"/>
+                            <input type="submit" class="btn btn-success" name="modify_book" value="Save Changes"/>
+                            <input type="submit" class="btn btn-danger" name="delete_book" value="Delete Book"/>
+                            <input type="submit" class="btn btn-primary" name="cancel" value="Cancel"/>
                             </form>
-                            <p>
-                                <a href="bookCatalogue.php"> Back to catalogue </a>
-                            </p>
-                        </body>
-                    </html>
-
-                <?php
+						</div>
+					</body>
+					</html>
+				 <?php
                         }
             }
         }
