+++ newUserReg.php
@@ -160,37 +160,116 @@ date_default_timezone_set("Asia/Calcutta");
 		}
 	}
 ?>
+
+
 <html>
 <head>
-<title> New User Registration </title>
+<title>Biblio@DSCE</title>
+	
+	<script src="https://kit.fontawesome.com/5d3eee0a99.js" crossorigin="anonymous"></script>
+	
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1">
+	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+	
+	
+	<link rel="stylesheet" type="text/css" href="homeu.css">
+	<style>
+		.page-content td{
+			vertical-align:middle;
+		}
+	</style>
 </head>
 <body>
-	<form action="" method="post">
-	<label> First Name : 
-	<input type="text" name="fname_txt"/>
-	<span class="invalid-feedback"><?php echo $fname_err; ?></span>
-	</label><br/>
-	<label> Last Name :
-	<input type="text" name="lname_txt"/>
-	<span class="invalid-feedback"><?php echo $lname_err; ?></span>
-	</label><br/>
-	<label> Email Address :
-	<input type="email" name="email_txt"/>
-	<span class="invalid-feedback"><?php echo $email_err; ?></span>
-	</label><br/>
-	<label> Phone Number :
-	<input type="number" name="phone_num_txt"/>
-	<span class="invalid-feedback"><?php echo $phone_err; ?></span>
-	</label><br/>
-	<label> Address :
-	<input type="text" name="address_text"/>
-	<span class="invalid-feedback"><?php echo $address_err; ?></span>
-	</label><br/>
-	<br/>
-	<input type="submit" name="new_user_create" value="Create User"/>
-	</form>
-	<p>
-		<a href="welcome_staff.php">Back</a>
-	</p>
+		<div class="vertical-nav bg-dark text-light" id="sidebar">
+			<div class="py-4 px-3 mb-4 bg-dark text-light">
+				<div class="media d-flex align-item-center">
+					<img src="avatar1.png" alt="user image" width="80" height="80" class="mr-3 rounded-circle img-thumbnail shadow-sm">
+					<div class="media-body">
+					<h4 class="mt-3"> <?php echo htmlspecialchars($_SESSION["username"]); ?> </h4>
+					<p class="font-weight-normal text-muted mb-0">ADMIN</p>
+					</div>
+				</div>
+			</div>
+			
+			<p class="text-gray font-weight-bold text-uppercase px-3 small pb-4 mb-0">Dashboard</p>
+			<ul class="nav flex-column bg-white mb-0">
+				<li class="nav-item">
+					<a href="welcome_staff.php" class="nav-link bg-dark text-light"><i class="fa fa-th-large mr-3 text-primary fa-fw"></i>home</a>
+				</li>
+				<li class="nav-item">
+					<a href="staffsearch.php" class="nav-link bg-dark text-light"><i class="fas fa-search mr-3 text-primary fa-fw"></i>search book</a>
+				</li>
+				<li class="nav-item">
+					<a href="newbook.php" class="nav-link active bg-dark text-light"><i class="fas fa-plus-square mr-3 text-primary fa-fw"></i>add book</a>
+				</li>
+				<li class="nav-item">
+					<a href="bookcatalogue.php" class="nav-link bg-dark text-light"><i class="fas fa-book-reader mr-3 text-primary fa-fw"></i>Book Catalogue</a>
+				</li>
+				<li class="nav-item">
+					<a href="staffborrow.php" class="nav-link bg-dark text-light"><i class="fas fa-book-open mr-3 text-primary fa-fw"></i>issue book</a>
+				</li>
+				<li class="nav-item">
+					<a href="outstandingpage.php" class="nav-link bg-dark text-light"><i class="fas fa-book mr-3 text-primary fa-fw"></i>outstanding books</a>
+				</li>
+				<li class="nav-item">
+					<a href="#" class="nav-link bg-dark text-dark"  id="highlight"><i class="fas fa-user-plus mr-3 text-primary fa-fw"></i>register new user</a>
+				</li>
+			</ul>
+			
+			<p class="text-gray font-weight-bold text-uppercase px-3 small py-4 mb-0">Charts</p>
+			<ul class="nav flex-column bg-white mb-0">
+				<li class="nav-item">
+					<a href="#" class="nav-link bg-dark text-light"><i class="fas fa-clipboard mr-3 text-primary fa-fw"></i>report</a>
+				</li>
+				<li class="nav-item">
+					<a href="#" class="nav-link bg-dark text-light"><i class="fas fa-chart-bar mr-3 text-primary fa-fw"></i>statistics</a>
+				</li>
+			</ul>	
+			
+			<p class="text-gray font-weight-bold text-uppercase px-3 small py-4 mb-0"></p>
+			<ul class="nav flex-column bg-white mb-0">
+				<li class="nav-item">
+					<a href="logout_staff.php" class="nav-link bg-dark text-light"><i class="fas fa-sign-out-alt mr-3 text-primary fa-fw"></i>logout</a>
+				</li>
+			</ul>
+		
+		</div>
+		
+		<div class="page-content p-5" id="content">
+			<h1>Register New User</h1><br/>
+			
+			<form action="" method="post">
+				<div class="row mb-3">
+				<label class="col-sm-2 col-form-label">First Name : </label>
+				<div class="col-sm-5"><input type="text" class="form-control" autocomplete="off" name="fname_txt"/></div>
+				<span class="invalid-feedback"><?php echo $fname_err; ?></span>
+				</div>
+				<div class="row mb-3">
+				<label class="col-sm-2 col-form-label">Last Name :</label>
+				<div class="col-sm-5"><input type="text" class="form-control" autocomplete="off" name="lname_txt"/></div>
+				<span class="invalid-feedback"><?php echo $lname_err; ?></span>
+				</div>
+				<div class="row mb-3">
+				<label class="col-sm-2 col-form-label">Email Address :</label>
+				<div class="col-sm-5"><input type="email" name="email_txt" autocomplete="off" class="form-control"/></div>
+				<span class="invalid-feedback"><?php echo $email_err; ?></span>
+				</div>
+				<div class="row mb-3">
+				<label class="col-sm-2 col-form-label"> Phone Number :</label>
+				<div class="col-sm-5"><input type="number" name="phone_num_txt" autocomplete="off" class="form-control"/></div>
+				<span class="invalid-feedback"><?php echo $phone_err; ?></span>
+				</div>
+				<div class="row mb-3">
+				<label class="col-sm-2 col-form-label"> Address : </label>
+				<div class="col-sm-5"><input type="text" class="form-control" autocomplete="off" name="address_text"/></div>
+				<span class="invalid-feedback"><?php echo $address_err; ?></span>
+				</div>
+				<br/>
+				<input type="submit" class="btn btn-primary" name="new_user_create" value="Create User"/>
+				</form>
+			
+			
+		</div>
 </body>
 </html>
