+++ outstandingPage.php
@@ -83,77 +83,152 @@ date_default_timezone_set("Asia/Calcutta");
 	$result = $link->query($outstand_displayQue);
 	
 	?>
+	
+	
 	<html>
-	<head>
-	<title> Outstanding Books </title>
-	</head>
-	<body>
-	<?php
-	if($result == True) {
-		$row_num = mysqli_num_rows($result);
-		
-		if($row_num>0) {
-			//display in html table
-			?>
-				<table border="1" cellpadding="10">
-					<tr>
-					<th> Title </th>
-					<th> Author </th>
-					<th> Issued by</th>
-					<th> Reader ID </th>
-					<th> Issue Date </th>
-					<th> Due Date </th>
-					<th> Return </th>
-					</tr>
-					<?php
-						while($row = $result->fetch_assoc())
-						{
-						?>	<tr>
-							<td><?php echo $row['title']?></td>
-							<td><?php echo $row['author']?></td>
-							<td><?php echo $row['name']?></td>
-							<td><?php echo $row['readerID']?></td>
-							<td><?php echo $row['issueDate']?></td>	
-							<td><?php echo $row['dueDate']?></td>	
-							<td>
-								<!-- Return button here -->
-								<form action="<?php echo htmlspecialchars($_SERVER["PHP_SELF"]);?>" method="post">
-								<button type="submit" value="<?php echo $row['bookID']?>" name="return_button"> Return </button>
-								</form>
-							</td>
-							<!--<td>
-							<form action="<?php //echo htmlspecialchars($_SERVER["PHP_SELF"]);?>" method="post">
-							
-							<?php 
-								// if the book was reserved, and is now available, we give a issue option
-								//if($row['availability'] == 'Available_Reserved') 
-								//{
-								?>
-									<button name="issue_reserve" value="<?php //echo $row['bookID'];?>" type="submit"> Issue </button>
-								<?php
-									
-								//}
-							
-							?>
-							
-							<button name="cancel_reserve" value="<?php //echo $row['bookID'];?>" type="submit"> Cancel </button>
-							</form>
-							</td>-->
-							</tr>
-						<?php
-						}
-					?>
-				</table>
-			<?php
-		} else {
-			echo "No outstanding books.";
+<head>
+<title>Biblio@DSCE</title>
+	
+	<script src="https://kit.fontawesome.com/5d3eee0a99.js" crossorigin="anonymous"></script>
+	
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1">
+	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+	
+	
+	<link rel="stylesheet" type="text/css" href="homeu.css">
+	<style>
+		.page-content td{
+			vertical-align:middle;
 		}
-	} else {
-		echo "No outstanding books.";
-	}
-?>	
-	<p>
-	<a href="welcome_staff.php">Back</a>
-	</p>
-	</body>
-	</html>
+	</style>
+</head>
+<body>
+		<div class="vertical-nav bg-dark text-light" id="sidebar">
+			<div class="py-4 px-3 mb-4 bg-dark text-light">
+				<div class="media d-flex align-item-center">
+					<img src="avatar1.png" alt="user image" width="80" height="80" class="mr-3 rounded-circle img-thumbnail shadow-sm">
+					<div class="media-body">
+					<h4 class="mt-3"> <?php echo htmlspecialchars($_SESSION["username"]); ?> </h4>
+					<p class="font-weight-normal text-muted mb-0">ADMIN</p>
+					</div>
+				</div>
+			</div>
+			
+			<p class="text-gray font-weight-bold text-uppercase px-3 small pb-4 mb-0">Dashboard</p>
+			<ul class="nav flex-column bg-white mb-0">
+				<li class="nav-item">
+					<a href="welcome_staff.php" class="nav-link bg-dark text-light"><i class="fa fa-th-large mr-3 text-primary fa-fw"></i>home</a>
+				</li>
+				<li class="nav-item">
+					<a href="staffsearch.php" class="nav-link bg-dark text-light"><i class="fas fa-search mr-3 text-primary fa-fw"></i>search book</a>
+				</li>
+				<li class="nav-item">
+					<a href="newbook.php" class="nav-link active bg-dark text-light"><i class="fas fa-plus-square mr-3 text-primary fa-fw"></i>add book</a>
+				</li>
+				<li class="nav-item">
+					<a href="bookcatalogue.php" class="nav-link bg-dark text-light"><i class="fas fa-book-reader mr-3 text-primary fa-fw"></i>Book Catalogue</a>
+				</li>
+				<li class="nav-item">
+					<a href="staffborrow.php" class="nav-link bg-dark text-light"><i class="fas fa-book-open mr-3 text-primary fa-fw"></i>issue book</a>
+				</li>
+				<li class="nav-item">
+					<a href="#" class="nav-link bg-dark text-dark"  id="highlight"><i class="fas fa-book mr-3 text-primary fa-fw"></i>outstanding books</a>
+				</li>
+				<li class="nav-item">
+					<a href="newUserReg.php" class="nav-link bg-dark text-light"><i class="fas fa-user-plus mr-3 text-primary fa-fw"></i>register new user</a>
+				</li>
+			</ul>
+			
+			<p class="text-gray font-weight-bold text-uppercase px-3 small py-4 mb-0">Charts</p>
+			<ul class="nav flex-column bg-white mb-0">
+				<li class="nav-item">
+					<a href="#" class="nav-link bg-dark text-light"><i class="fas fa-clipboard mr-3 text-primary fa-fw"></i>report</a>
+				</li>
+				<li class="nav-item">
+					<a href="#" class="nav-link bg-dark text-light"><i class="fas fa-chart-bar mr-3 text-primary fa-fw"></i>statistics</a>
+				</li>
+			</ul>	
+			
+			<p class="text-gray font-weight-bold text-uppercase px-3 small py-4 mb-0"></p>
+			<ul class="nav flex-column bg-white mb-0">
+				<li class="nav-item">
+					<a href="logout_staff.php" class="nav-link bg-dark text-light"><i class="fas fa-sign-out-alt mr-3 text-primary fa-fw"></i>logout</a>
+				</li>
+			</ul>
+		
+		</div>
+		
+		<div class="page-content p-5" id="content">
+			<h1>Outstanding Books</h1><br/>
+			
+			<?php
+				if($result == True) {
+					$row_num = mysqli_num_rows($result);
+					
+					if($row_num>0) {
+						//display in html table
+						?>
+							<table class="table table-striped table-bordered">
+								<thead class="thead-dark">
+								<tr>
+								<th> Title </th>
+								<th> Author </th>
+								<th> Issued by</th>
+								<th> Reader ID </th>
+								<th> Issue Date </th>
+								<th> Due Date </th>
+								<th> Return </th>
+								</tr></thead>
+								<?php
+									while($row = $result->fetch_assoc())
+									{
+									?>	<tr>
+										<td><?php echo $row['title']?></td>
+										<td><?php echo $row['author']?></td>
+										<td><?php echo $row['name']?></td>
+										<td><?php echo $row['readerID']?></td>
+										<td><?php echo $row['issueDate']?></td>	
+										<td><?php echo $row['dueDate']?></td>	
+										<td>
+											<!-- Return button here -->
+											<form action="<?php echo htmlspecialchars($_SERVER["PHP_SELF"]);?>" method="post">
+											<button type="submit" class="btn btn-primary" value="<?php echo $row['bookID']?>" name="return_button"> Return </button>
+											</form>
+										</td>
+										<!--<td>
+										<form action="<?php //echo htmlspecialchars($_SERVER["PHP_SELF"]);?>" method="post">
+										
+										<?php 
+											// if the book was reserved, and is now available, we give a issue option
+											//if($row['availability'] == 'Available_Reserved') 
+											//{
+											?>
+												<button name="issue_reserve" value="<?php //echo $row['bookID'];?>" type="submit"> Issue </button>
+											<?php
+												
+											//}
+										
+										?>
+										
+										<button name="cancel_reserve" value="<?php //echo $row['bookID'];?>" type="submit"> Cancel </button>
+										</form>
+										</td>-->
+										</tr>
+									<?php
+									}
+								?>
+							</table>
+						<?php
+					} else {
+						echo "No outstanding books.";
+					}
+				} else {
+					echo "No outstanding books.";
+				}
+			?>
+		
+		
+		</div>
+</body>
+</html>
